DEV-12829 Further refinement
